Adds a missing include, over the shoulder by @larkost

+++ src/unittest/timer_test.cc
@@ -1,6 +1,8 @@
 // Copyright 2010-2014 RethinkDB, all rights reserved.
 #include "unittest/gtest.hpp"
 
+#include <algorithm>
+
 #include "arch/timing.hpp"
 #include "concurrency/pmap.hpp"
 #include "unittest/unittest_utils.hpp"
@@ -19,7 +21,9 @@ void walk_wait_times(int i) {
         const int64_t diff = static_cast<int64_t>(t2) - static_cast<int64_t>(t);
         // Asserts that we're off by less than 25% of the sleep time
         // or two milliseconds, whichever is larger.
-        ASSERT_LT(llabs(diff - wait_array[i][j] * MILLION), std::max(diff / 4, 2 * MILLION));
+        ASSERT_LT(
+            llabs(diff - wait_array[i][j] * MILLION),
+            std::max(diff / 4, static_cast<int64_t>(2 * MILLION)));
         t = t2;
     }
